atmospheric_harvester/requirements.txt
pygame-ce
aiohttp
